Reuse connection for subsequent API requests
Use a session so that requests to the Recurse Center API reuse the
connection created by the first request, rather than set up a new
connection for each request. This should both improve performance of the
update and reduce the load imposed on recurse.com.

This can be seen by setting the log level to DEBUG and watching for
lines like this:

    DEBUG:urllib3.connectionpool:Starting new HTTPS connection (1): www.recurse.com:443

After this change, there should only be one such log line.

See also:
https://2.python-requests.org//en/latest/user/advanced/#keep-alive