Always show footer, even on mobile
While we might want to hide the footer on mobile, we'd only want to do
that if we can take advantage of the space effectively by
enlarging the map by the corresponding amount when we hide the
footer. Until we do that, show the footer on mobile by removing both
display-related bootstrap classes[1]: the one that hides it, and the one
that overrides the hiding when the screen width is large enough.

[1] https://getbootstrap.com/docs/4.3/utilities/display/